Move LHHeadAssembly into it's own file
